Move logic to a better named component
